Fundamentals/code_clinic/jhb19/requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
prettyTable
colorama
click